add docs to disk.py functions

+++ invisibleroads_macros/disk.py
@@ -2,9 +2,11 @@ import fnmatch
 import re
 import tarfile
 from contextlib import contextmanager
-from os import chdir, getcwd, makedirs, remove, walk
+from os import chdir, getcwd, makedirs, remove, walk, listdir
 from os.path import (
-    abspath, basename, dirname, exists, isfile, join, normpath, realpath,
+    abspath, basename, dirname,
+    exists, isdir, isfile,
+    join, normpath, realpath,
     relpath, sep, splitext)
 from pathlib import Path
 from shutil import copytree, rmtree
@@ -13,6 +15,9 @@ from zipfile import ZipFile, ZIP_DEFLATED
 
 
 def make_folder(folder):
+    """
+        makes a folder, doesn't raise an exception if folder already exists
+    """
     try:
         makedirs(folder)
     except OSError:
@@ -21,18 +26,27 @@ def make_folder(folder):
 
 
 def clean_folder(folder):
+    """
+        removes folder contents
+    """
     remove_path(folder)
     return make_folder(folder)
 
 
+# TODO: check if target_folder is dir?
 def replace_folder(target_folder, source_folder):
+    """
+        replaces a folder with source_folder
+    """
+    if isfile(target_folder):
+        raise OSError('must pass directory')
     remove_path(target_folder)
-    make_folder(dirname(target_folder))
     copytree(source_folder, target_folder)
     return target_folder
 
 
 def remove_path(path):
+    """ removes a file or directory from disk """
     try:
         rmtree(path)
     except OSError:
@@ -44,36 +58,54 @@ def remove_path(path):
 
 
 def get_nickname(path):
+    """
+        returns the name of file,
+        example: ./file.txt -> file
+    """
     return splitext(basename(path))[0]
 
 
 def make_link(source_path, target_path):
-    from os import readlink, symlink
+    """
+        creates a symbolic link to source_path
+    """
+    from os import symlink
     source_path = normpath(source_path)
     if not exists(target_path):
         symlink(source_path, target_path)
         return target_path
-    try:
-        if normpath(readlink(target_path)) != source_path:
-            raise IOError('could not make link; target_path is another link')
-    except OSError:
-        path_type = 'file' if isfile(target_path) else 'folder'
-        raise IOError('could not make link; target_path is a %s' % path_type)
+    else:
+        raise OSError('{0}: file exists'.format(target_path))
     return target_path
 
 
-def find_path(name, folder):
-    for root_folder, folder_names, file_names in walk(folder):
-        if name in file_names:
-            return join(root_folder, name)
+def find_path(name, folder, recursive=True):
+    """
+        finds the file in folder/sub-folders
+    """
+    if recursive:
+        for root_folder, folder_names, file_names in walk(folder):
+            if name in file_names:
+                return join(root_folder, name)
+    else:
+        if name in listdir(folder):
+            return join(folder, name)
     raise IOError('cannot find "%s" in "%s"' % (name, folder))
 
 
-def find_paths(name_expression, folder):
-    return [
-        join(root_folder, file_name)
-        for root_folder, folder_names, file_names in walk(folder)
-        for file_name in fnmatch.filter(file_names, name_expression)]
+def find_paths(name_expression, folder, recursive=True):
+    """
+        finds files matching the expression in folder,
+        and subfolders if recursive is specified
+    """
+    if recursive:
+        return [
+            join(root_folder, file_name)
+            for root_folder, folder_names, file_names in walk(folder)
+            for file_name in fnmatch.filter(file_names, name_expression)]
+    return [join(folder, filename) for filename in listdir(folder) if isfile(
+            join(folder, filename)) and fnmatch.fnmatch(
+                filename, name_expression)]
 
 
 def resolve_relative_path(relative_path, folder):
@@ -84,6 +116,11 @@ def resolve_relative_path(relative_path, folder):
 
 
 def compress(source_folder, target_path=None):
+    """
+        compresses a folder to a target_path
+        if folder ends with '.tar.gz' a tarball is produced
+        otherwise it is compressed into a zip file
+    """
     if not target_path:
         target_path = source_folder + '.tar.gz'
     if target_path.endswith('.tar.gz'):
@@ -94,13 +131,17 @@ def compress(source_folder, target_path=None):
 
 
 def compress_tar_gz(source_folder, target_path=None):
+    """
+        compresses into a tarball
+    """
     if not target_path:
         target_path = source_folder + '.tar.gz'
     with tarfile.open(target_path, 'w:gz', dereference=True) as target_file:
-        for path in Path(source_folder).rglob('*'):
-            if path.is_dir():
+        for path in find_paths('*', source_folder):
+            # TODO: add directories
+            if isdir(path):
                 continue
-            target_file.add(str(path), str(path.relative_to(source_folder)))
+            target_file.add(str(path), str(relpath(path, source_folder)))
     return target_path
